fix: resolve 500 errors in merchant order API
- Fix orderId type mismatch (Long to Integer)
- Add null check for order items in getMerchantOrders
- Improve isOrderBelongToMerchant to check all items

+++ backend/src/main/java/com/healthmall/controller/MerchantOrderController.java
@@ -53,7 +53,7 @@ public class MerchantOrderController {
 
     @PostMapping("/{orderId}/confirm")
     public ApiResponse<OrderResponse> confirmOrder(
-            @PathVariable Long orderId,
+            @PathVariable Integer orderId,
             HttpServletRequest request) {
         Integer merchantId = (Integer) request.getAttribute("userId");
         if (merchantId == null) {
@@ -66,7 +66,7 @@ public class MerchantOrderController {
 
     @PostMapping("/{orderId}/reject")
     public ApiResponse<OrderResponse> rejectOrder(
-            @PathVariable Long orderId,
+            @PathVariable Integer orderId,
             @RequestBody ConfirmOrderRequest request,
             HttpServletRequest request) {
         Integer merchantId = (Integer) request.getAttribute("userId");
